factor out candidate selection to its own set of pipeline components, update component factory decorators to Language.factory for spaCy v3 compat
